refactored model components
